add link to tarball; enhance rsync script

+++ www/trunk/xml_export/rsync.sh
@@ -1,5 +1,33 @@
 #!/bin/bash
-rm res/Export.tgz
-tar czf res/Export.tgz --exclude=Export.tgz res
-cd res
+# push the transformed tracker export to SF hosting
+#
+# Copyright © 2012 - Eric Le Lay <kerik-sf@users.sf.net>
+# This program is free software; you can redistribute it and/or
+# modify it under the terms of the GNU General Public License
+# as published by the Free Software Foundation; either version 2
+# of the License, or any later version.
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+# 
+# You should have received a copy of the GNU General Public License
+# along with this program; if not, write to the Free Software
+# Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+#/
+usage="
+Usage: $0 DEST_DIR
+
+    DEST_DIR  : directory where index.html and subdirs per tracker are
+"
+
+if [ $# != 1 ]; then
+	echo "$usage"
+	exit -1
+fi
+
+destdir=$1
+rm "$destdir/Export.tgz"
+tar czf "$destdir/Export.tgz" --exclude=Export.tgz "$destdir"
+cd "$destdir"
 rsync -avP -e ssh --delete . kerik-sf@web.sourceforge.net:/home/project-web/jedit/htdocs/trackers
